airframe-control: Support Scala.js
