+++ airframe-control/src/test/scala/wvlet/airframe/control/CircuitBreakerTest.scala
@@ -4,6 +4,7 @@ import wvlet.airspec._
 import java.util.concurrent.TimeoutException
 
 class CircuitBreakerTest extends AirSpec {
+  scalaJsSupport
 
   def `support changing states`: Unit = {
     val cb = CircuitBreaker.default
@@ -104,33 +105,6 @@ class CircuitBreakerTest extends AirSpec {
     cb.halfOpen
 
     // 1/4
-    cb.recordSuccess
-    cb.isConnected shouldBe true
-  }
-
-  def `support failure rate health checker`: Unit = {
-    val cb = CircuitBreaker.withFailureRate(0.01, timeWindowMillis = 1000)
-    val e  = new TimeoutException()
-    cb.isConnected shouldBe true
-
-    // 1/1
-    cb.recordSuccess
-    cb.isConnected shouldBe true
-
-    // 1/2
-    Thread.sleep(200)
-    cb.recordFailure(e)
-    cb.isConnected shouldBe false
-
-    // 1/3
-    Thread.sleep(200)
-    cb.recordFailure(e)
-    cb.isConnected shouldBe false
-
-    // Force probing
-    cb.halfOpen
-
-    // The state should be recovered after the successful request
     cb.recordSuccess
     cb.isConnected shouldBe true
   }
