+++ airframe-control/src/test/scala/wvlet/airframe/control/RetryTest.scala
@@ -22,6 +22,8 @@ import scala.concurrent.TimeoutException
   *
   */
 class RetryTest extends AirSpec {
+  scalaJsSupport
+
   def `support backoff retry`: Unit = {
     var count = 0
